pyproject.toml
[project]
name = "excel-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "dibkb",email = "dibas9110@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "langchain (>=0.3.19,<0.4.0)",
    "langchain-openai (>=0.3.6,<0.4.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "langchain-groq (>=0.2.4,<0.3.0)",
    "python-socketio (>=5.12.1,<6.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "dibkb-scraper (==0.2.7)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "src" }
]

[tool.poetry.scripts]
dev = "src.run:main"
